software/software/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect2==0.7.0
cx-Oracle==8.3.0
Django==3.2.25
djangorestframework==3.15.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
h11==0.14.0
html5lib==1.1
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
lxml==5.3.0
openpyxl==3.1.3
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.0
Pillow==9.5.0
pluggy==1.2.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyMySQL==1.1.1
pypdf==3.1.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.4.4
python-bidi==0.6.0
python-dotenv==0.21.1
pytz==2024.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2024.2
tzlocal==5.1
uritools==4.0.3
urllib3==1.26.15
webdriver-manager==4.0.2
webencodings==0.5.1
wsproto==1.2.0
xhtml2pdf==0.2.11
XlsxWriter==3.2.0
zipp==3.15.0
